Created grid and added state.
I componentised the title(Header) and the body(PhotoGrid) which are both imported into App. I created an example state object to look through an array of numbers which I then passed to PhotoGrid. I assigned a key to each item which then prints each number w/in a div to each grid square.

Next, I plan to fetch images from the unsplash API which I will store x amount of images in state and pass them as props to PhotoGrid.